+++ app/src/common/apiReference/TypeRefSection.module.scss
@@ -2,13 +2,6 @@
     max-width: 100%;
     width: 960px;
 
-    .sticky-header {
-        position: sticky;
-        top: 0;
-        isolation: isolate;
-        z-index: 5;
-    }
-
     :global(.asterisk) {
         color: var(--uui-labeled-input-asterisk);
         font-family: var(--font-semibold);
